+++ command/composer/composer.go
@@ -1,19 +1,28 @@
 package composer
 
 import (
+	"bytes"
 	"fmt"
+	"os"
 	"os/exec"
+	"path/filepath"
+	"strings"
 
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/filters"
+	"github.com/docker/docker/api/types/network"
 	"github.com/docker/docker/client"
+	"github.com/docker/docker/pkg/stdcopy"
 	"github.com/spf13/cobra"
 
+	"github.com/craftcms/nitro/pkg/composer"
 	"github.com/craftcms/nitro/pkg/config"
 	"github.com/craftcms/nitro/pkg/containerargs"
 	"github.com/craftcms/nitro/pkg/containerlabels"
 	"github.com/craftcms/nitro/pkg/contextor"
 	"github.com/craftcms/nitro/pkg/terminal"
+	"github.com/craftcms/nitro/pkg/volumename"
+	volumetypes "github.com/docker/docker/api/types/volume"
 )
 
 var (
@@ -79,6 +88,7 @@ func NewCommand(home string, docker client.CommonAPIClient, output terminal.Outp
 				return err
 			}
 
+			// if the container was defined
 			if command.Container != "" {
 				// make sure its a valid site
 				if _, err := cfg.FindSiteByHostName(command.Container); err != nil {
@@ -131,6 +141,151 @@ func NewCommand(home string, docker client.CommonAPIClient, output terminal.Outp
 				output.Info("composer", command.Args[0], "completed 🤘")
 
 				return nil
+			}
+
+			// fallback to the stand alone container
+			image := "docker.io/composer"
+
+			// filter for the image ref
+			filter := filters.NewArgs()
+			filter.Add("reference", image)
+
+			// look for the image
+			images, err := docker.ImageList(ctx, types.ImageListOptions{Filters: filter})
+			if err != nil {
+				return fmt.Errorf("unable to get a list of images, %w", err)
+			}
+
+			// if we don't have the image, pull it
+			if len(images) == 0 {
+				rdr, err := docker.ImagePull(ctx, image, types.ImagePullOptions{All: false})
+				if err != nil {
+					return fmt.Errorf("unable to pull the docker image, %w", err)
+				}
+
+				buf := &bytes.Buffer{}
+				if _, err := buf.ReadFrom(rdr); err != nil {
+					return fmt.Errorf("unable to read the output from pulling the image, %w", err)
+				}
+			}
+
+			// remove the image ref filter
+			filter.Del("reference", image)
+
+			// find the network
+			networkFilter := filters.NewArgs()
+			networkFilter.Add("name", "nitro-network")
+
+			// check if the network needs to be created
+			networks, err := docker.NetworkList(ctx, types.NetworkListOptions{Filters: networkFilter})
+			if err != nil {
+				return fmt.Errorf("unable to list the docker networks, %w", err)
+			}
+
+			var networkID string
+			for _, n := range networks {
+				if n.Name == "nitro-network" || strings.TrimLeft(n.Name, "/") == "nitro-network" {
+					networkID = n.ID
+				}
+			}
+
+			wd, err := os.Getwd()
+			if err != nil {
+				return fmt.Errorf("unable to get the current directory, %w", err)
+			}
+
+			path, err := filepath.Abs(wd)
+			if err != nil {
+				return fmt.Errorf("unable to find the absolute path, %w", err)
+			}
+
+			// add filters for the volume
+			filter.Add("label", containerlabels.Type+"=composer")
+			filter.Add("label", containerlabels.Path+"="+path)
+
+			// check if there is an existing volume
+			volumes, err := docker.VolumeList(ctx, filter)
+			if err != nil {
+				return err
+			}
+
+			// set the volume name
+			volumeName := volumename.FromPath(path)
+
+			var pathVolume types.Volume
+			switch len(volumes.Volumes) {
+			case 1:
+				pathVolume = *volumes.Volumes[0]
+			case 0:
+				// create the volume if it does not exist
+				volume, err := docker.VolumeCreate(ctx, volumetypes.VolumeCreateBody{
+					Driver: "local",
+					Name:   volumeName,
+					Labels: map[string]string{
+						containerlabels.Type: "composer",
+						containerlabels.Path: path,
+					},
+				})
+				if err != nil {
+					return fmt.Errorf("unable to create the volume, %w", err)
+				}
+
+				pathVolume = volume
+			}
+
+			// build the container options
+			opts := &composer.Options{
+				Image:    image,
+				Commands: args,
+				Labels: map[string]string{
+					containerlabels.Nitro: "true",
+					containerlabels.Type:  "composer",
+					containerlabels.Path:  path,
+				},
+				Volume: &pathVolume,
+				Path:   path,
+				NetworkConfig: &network.NetworkingConfig{
+					EndpointsConfig: map[string]*network.EndpointSettings{
+						"nitro-network": {
+							NetworkID: networkID,
+						},
+					},
+				},
+			}
+
+			// create the container
+			container, err := composer.CreateContainer(ctx, docker, opts)
+			if err != nil {
+				return fmt.Errorf("unable to create the composer container\n%w", err)
+			}
+
+			// attach to the container
+			stream, err := docker.ContainerAttach(ctx, container.ID, types.ContainerAttachOptions{
+				Stream: true,
+				Stdout: true,
+				Stderr: true,
+				Logs:   true,
+			})
+			if err != nil {
+				return fmt.Errorf("unable to attach to container, %w", err)
+			}
+			defer stream.Close()
+
+			// run the container
+			if err := docker.ContainerStart(ctx, container.ID, types.ContainerStartOptions{}); err != nil {
+				return fmt.Errorf("unable to start the container, %w", err)
+			}
+
+			// show the output to stdout and stderr
+			if _, err := stdcopy.StdCopy(os.Stdout, os.Stderr, stream.Reader); err != nil {
+				return fmt.Errorf("unable to copy the output of the container logs, %w", err)
+			}
+
+			output.Info("composer command completed 🤘")
+
+			// remove the container
+			if err := docker.ContainerRemove(ctx, container.ID, types.ContainerRemoveOptions{}); err != nil {
+				return err
 			}
 
 			return nil